SVD and ALS models for Collaborative Filtering
This notebook contains implementations of matrix factorisation based algorithms for collaborative filtering. It includes the SVD and Baseline ALS models to make video game recommendations to users of the Playstation 4 as well as explanations on the implementations at different stages of the process.

The models have been built using the Python Surprise Package.