MLP and NeuMF Keras Models for PSN RecSys
This notebook contains an analysis of the Multi-Layer Perceptron and Neural Matrix Factorization models for collaborative filtering implemented using keras. The model has been build in the keras_models/optimization folder and this notebook does not deal with the actual model creation. However, it does go into detail on the model architecture.