MLP and NeuMF Keras Models for PSN RecSys
This notebook contains an analysis of the Multi-Layer Perceptron and Neural Matrix Factorization models for collaborative filtering implemented using keras. The model has been built and optimized in the Keras_Models/Hyperparameter_Optim folder and this notebook does not deal with the actual model creation. However, it does go into detail on the model architecture.